+++ autoplex/data/common/jobs.py
@@ -594,47 +594,39 @@ def collect_dft_data(
 
     for i, val in enumerate(dirs):
         if os.path.exists(os.path.join(val, "vasprun.xml.gz")):
-            try:
-                converged = check_convergence_vasp(os.path.join(val, "vasprun.xml.gz"))
+            at = read(os.path.join(val, "vasprun.xml.gz"), index=":")
+            for at_i in at:
+                virial_list = (
+                    -voigt_6_to_full_3x3_stress(at_i.get_stress())
+                    * at_i.get_volume()
+                )
+                at_i.info["REF_virial"] = " ".join(
+                    map(str, virial_list.flatten())
+                )
+                del at_i.calc.results["stress"]
+                at_i.arrays["REF_forces"] = at_i.calc.results["forces"]
+                del at_i.calc.results["forces"]
+                at_i.info["REF_energy"] = at_i.calc.results["free_energy"]
+                del at_i.calc.results["energy"]
+                del at_i.calc.results["free_energy"]
+                atoms.append(at_i)
+                at_i.info["config_type"] = config_types[i]
+                if (
+                    at_i.info["config_type"] != "dimer"
+                    and at_i.info["config_type"] != "IsolatedAtom"
+                ):
+                    at_i.pbc = True
+                    at_i.info["rss_group"] = rss_group
+                else:
+                    at_i.info["rss_nonperiodic"] = "T"
 
-                if converged:
-                    at = read(os.path.join(val, "vasprun.xml.gz"), index=":")
-                    for at_i in at:
-                        virial_list = (
-                            -voigt_6_to_full_3x3_stress(at_i.get_stress())
-                            * at_i.get_volume()
-                        )
-                        at_i.info["REF_virial"] = " ".join(
-                            map(str, virial_list.flatten())
-                        )
-                        del at_i.calc.results["stress"]
-                        at_i.arrays["REF_forces"] = at_i.calc.results["forces"]
-                        del at_i.calc.results["forces"]
-                        at_i.info["REF_energy"] = at_i.calc.results["free_energy"]
-                        del at_i.calc.results["energy"]
-                        del at_i.calc.results["free_energy"]
-                        atoms.append(at_i)
-                        at_i.info["config_type"] = config_types[i]
-                        if (
-                            at_i.info["config_type"] != "dimer"
-                            and at_i.info["config_type"] != "IsolatedAtom"
-                        ):
-                            at_i.pbc = True
-                            at_i.info["rss_group"] = rss_group
-                        else:
-                            at_i.info["rss_nonperiodic"] = "T"
+                if at_i.info["config_type"] == "IsolatedAtom":
+                    at_ids = at_i.get_atomic_numbers()
+                    # array_key = at_ids.tostring()
+                    isolated_atom_energies[int(at_ids[0])] = at_i.info[
+                        "REF_energy"
+                    ]
 
-                        if at_i.info["config_type"] == "IsolatedAtom":
-                            at_ids = at_i.get_atomic_numbers()
-                            # array_key = at_ids.tostring()
-                            isolated_atom_energies[int(at_ids[0])] = at_i.info[
-                                "REF_energy"
-                            ]
-
-            except ValueError:
-                logging.error(f"Failed to collect number: {i}")
-                failed_count += 1
-                traceback.print_exc()
 
     logging.info(f"Total {len(atoms)} structures from VASP are exactly collected.")
 
@@ -648,27 +640,6 @@ def collect_dft_data(
         "vasp_ref_dir": vasp_ref_dir,
         "isolated_atom_energies": isolated_atom_energies,
     }
-
-
-def check_convergence_vasp(path: str) -> bool:
-    """
-    Check if VASP calculation has converged.
-
-    Parameters
-    ----------
-    path: Path
-        Path to the vasp output file to check convergence.
-
-    Return
-    ------
-    bool
-        True if a run is converged both ionically and electronically.
-    """
-    vasprun = Vasprun(path)
-    converged_e = vasprun.converged_electronic
-    converged_i = vasprun.converged_ionic
-
-    return converged_e and converged_i
 
 
 def safe_strip_hostname(value):
