+++ tests/data/test_datagen_flows.py
@@ -13,6 +13,8 @@ from autoplex.data.phonons.flows import IsoAtomMaker, RandomStructuresDataGenera
 from atomate2.vasp.jobs.core import TightRelaxMaker
 from atomate2.vasp.sets.core import TightRelaxSetGenerator
 import pytest
+from jobflow import run_locally, Flow
+
 
 @pytest.fixture(scope="class")
 def relax_maker():
@@ -40,7 +42,6 @@ def relax_maker():
         )
 
 def test_ml_phonon_maker(test_dir, clean_dir, memory_jobstore):
-    from jobflow import run_locally
 
     potential_file = test_dir / "fitting" / "ref_files" / "gap_file.xml"
     path_to_struct = test_dir / "fitting" / "ref_files" / "POSCAR"
@@ -68,7 +69,6 @@ def test_ml_phonon_maker(test_dir, clean_dir, memory_jobstore):
 
 
 def test_data_generation_distort_type_0(vasp_test_dir, mock_vasp, relax_maker, clean_dir):
-    from jobflow import run_locally
 
     path_to_struct = vasp_test_dir / "dft_ml_data_generation" / "POSCAR"
     structure = Structure.from_file(path_to_struct)
@@ -109,7 +109,6 @@ def test_data_generation_distort_type_0(vasp_test_dir, mock_vasp, relax_maker, c
 
 
 def test_data_generation_distort_type_1(vasp_test_dir, mock_vasp, relax_maker, clean_dir):
-    from jobflow import run_locally
 
     path_to_struct = vasp_test_dir / "dft_ml_data_generation" / "POSCAR"
     structure = Structure.from_file(path_to_struct)
@@ -197,7 +196,6 @@ def test_data_generation_distort_type_1(vasp_test_dir, mock_vasp, relax_maker, c
 
 
 def test_data_generation_distort_type_2(vasp_test_dir, mock_vasp, relax_maker, clean_dir):
-    from jobflow import run_locally
 
     path_to_struct = vasp_test_dir / "dft_ml_data_generation" / "POSCAR"
     structure = Structure.from_file(path_to_struct)
@@ -247,7 +245,6 @@ def test_data_generation_distort_type_2(vasp_test_dir, mock_vasp, relax_maker, c
 
 
 def test_data_generation_volume_range(vasp_test_dir, mock_vasp, relax_maker, clean_dir):
-    from jobflow import run_locally
 
     path_to_struct = vasp_test_dir / "dft_ml_data_generation" / "POSCAR"
     structure = Structure.from_file(path_to_struct)
@@ -308,7 +305,6 @@ def test_data_generation_volume_range(vasp_test_dir, mock_vasp, relax_maker, cle
 
 
 def test_iso_atom_maker(mock_vasp, clean_dir):
-    from jobflow import run_locally
     from pymatgen.core import Species
 
     specie = Species("Cl")
@@ -341,7 +337,6 @@ def test_iso_atom_maker(mock_vasp, clean_dir):
 def test_generate_training_data_for_testing(
         vasp_test_dir, test_dir, memory_jobstore, clean_dir
 ):
-    from jobflow import run_locally
 
     path_to_struct = vasp_test_dir / "dft_ml_data_generation" / "POSCAR"
     potential_file_dir = test_dir / "fitting" / "ref_files" / "gap_file.xml"
@@ -357,3 +352,96 @@ def test_generate_training_data_for_testing(
     responses = run_locally(
         generate_data, create_folders=True, ensure_success=False, store=memory_jobstore
     )  # atomate2 switched from pckl to json files for the trajectories --> job fails in its current state
+
+
+def test_vasp_static(test_dir, mock_vasp, memory_jobstore, clean_dir):
+    from autoplex.data.common.jobs import collect_dft_data
+    from autoplex.data.common.flows import DFTStaticLabelling
+    from ase.io import read
+    from pymatgen.core.structure import Structure
+    from atomate2.settings import Atomate2Settings
+    settings = Atomate2Settings()
+    
+    poscar_paths = {
+        f"static_bulk_{i}": test_dir / f"vasp/rss/Si_bulk_{i+1}/inputs/POSCAR"
+        for i in range(18)
+    }
+
+    test_structures = []
+    for path in poscar_paths.values():
+        structure = Structure.from_file(path)
+        test_structures.append(structure)
+
+    ref_paths = {
+        **{f"static_bulk_{i}": f"rss/Si_bulk_{i+1}/" for i in range(18)},
+        "static_isolated_0": "rss/Si_isolated_1/",
+        "static_dimer_0": "rss/Si_dimer_1/",
+        "static_dimer_1": "rss/Si_dimer_2/",
+        "static_dimer_2": "rss/Si_dimer_3/",
+    }
+
+    fake_run_vasp_kwargs = {
+        "static_isolated_0": {"incar_settings": {"ISPIN": 2, "KSPACINGS": 2.0}}, 
+        "static_dimer_0": {"incar_settings": {"ISPIN": 2, "KSPACINGS": 2.0}}, 
+        "static_dimer_1": {"incar_settings": {"ISPIN": 2, "KSPACINGS": 2.0}}, 
+        "static_dimer_2": {"incar_settings": {"ISPIN": 2, "KSPACINGS": 2.0}}, 
+    }
+
+    mock_vasp(ref_paths, fake_run_vasp_kwargs)
+
+    job1 = DFTStaticLabelling(isolated_atom=True, 
+                          e0_spin=True, 
+                          isolatedatom_box=[20.0, 20.5, 21.0],
+                          dimer=True, 
+                          dimer_box=[15.0, 15.5, 16.0],
+                          dimer_range=[1.5, 2.0],
+                          dimer_num=3,
+                          custom_incar={
+                            "ADDGRID": None, 
+                            "ENCUT": 200,
+                            "EDIFF": 1E-04,
+                            "ISMEAR": 0,
+                            "SIGMA": 0.05,
+                            "PREC": "Normal",
+                            "ISYM": None,
+                            "KSPACING": 0.3,
+                            "NPAR": 8,
+                            "LWAVE": "False",
+                            "LCHARG": "False",
+                            "ENAUG": None,
+                            "GGA": None,
+                            "ISPIN": None,
+                            "LAECHG": None,
+                            "LELF": None,
+                            "LORBIT": None,
+                            "LVTOT": None,
+                            "NSW": None,
+                            "SYMPREC": None,
+                            "NELM": 50,
+                            "LMAXMIX": None,
+                            "LASPH": None,
+                            "AMIN": None,
+                    },
+                    ).make(structures=test_structures)
+    
+    job2 = collect_dft_data(vasp_dirs=job1.output)
+    
+    response = run_locally(
+        Flow([job1,job2]),
+        create_folders=True,
+        ensure_success=True,
+        store=memory_jobstore
+    )
+
+    dict_vasp = job2.output.resolve(memory_jobstore)
+
+    path_to_vasp, isol_energy = dict_vasp['vasp_ref_dir'], dict_vasp['isolated_atom_energies']
+
+    atoms = read(path_to_vasp, index=":")
+    config_types = [at.info['config_type'] for at in atoms]
+    
+    assert isol_energy['14'] == -0.84696938
+    assert len(config_types) == 22
+    assert 'IsolatedAtom' in config_types
+    assert config_types.count("dimer") == 3
+    assert config_types.count("bulk") == 18
